move printBindings out of /config/index.ts
